requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
python-dotenv==0.17.1
python-http-client==3.3.2
sendgrid==6.7.0
six==1.16.0
SQLAlchemy==1.4.15
starkbank-ecdsa==1.1.0
virtualenv==20.4.6
Werkzeug==1.0.1
